feat: Фаза 6 — Super-Admin God Mode + бизнес логик засварууд
Super-admin бүх дэлгүүрийн мэдээллийг read-only горимоор харах боломжтой
болсон. Мөн борлуулалт, тооцоо, sync зэрэг бизнес логикийн алдаануудыг
засварласан.

Шинэ:
- GET /admin/stores/dashboard endpoint (бүх дэлгүүрийн нэгдсэн dashboard)
- AdminBrowseStore provider (browse mode state удирдлага)
- effectiveStoreId + isReadOnlyMode providers
- Super-admin dashboard: бүх салбарын борлуулалт, ашиг харуулах
- Browse mode: дэлгүүр дарж бараа, түүхийг read-only харах
- MainShell: browse mode banner + динамик navigation tabs
- Route guard: write routes хориглох, read routes нээх
- Write товчнууд нуугдах (FAB, Засах, Үлдэгдэл засах)

Засвар:
- cartTotal double→int шилжүүлт (бүх provider, screen)
- Auth guard route path зөрүү (7 route засварласан)
- Sales service race condition + cashTendered validation
- Shift service duplicate clock-in хамгаалалт
- Sync queue manager error handling сайжруулалт

+++ lib/features/sales/presentation/providers/cart_provider.dart
@@ -84,9 +84,9 @@ class CartNotifier extends _$CartNotifier {
 
 /// Cart total (sum of all item subtotals) - хөнгөлөлтийн дараах
 @riverpod
-double cartTotal(CartTotalRef ref) {
+int cartTotal(CartTotalRef ref) {
   final cartItems = ref.watch(cartNotifierProvider);
-  return cartItems.fold(0.0, (sum, item) => sum + item.subtotal);
+  return cartItems.fold(0, (sum, item) => sum + item.subtotal);
 }
 
 /// Нийт хөнгөлөлтийн дүн
